actualización archivos

+++ Clase5_Automovil/main.cpp
@@ -10,17 +10,17 @@ using namespace std;
 
 int main(){
     cout<<"\nSistema de control de uso de autos\n"; 
-    Automovil miCoche;
-    miCoche.imprimeAutomovil();
-    cout<<"\nQue distancia quieres recorrer?: ";
-    float distancia;
-    cin>>distancia;
-    cout<<"Cuantos pasajeros viajaran?: ";
-    int pasajeros;
+    Automovil miCoche; 
+    miCoche.imprimeAutomovil(); 
+    cout<<"\nQue distancia quieres recorrer?: "; 
+    float distancia; 
+    cin>>distancia; 
+    cout<<"Cuantos pasajeros viajaran?: "; 
+    float pasajeros; 
     cin>>pasajeros; 
     miCoche.avanzaAutomovil(distancia,pasajeros); 
     miCoche.imprimeAutomovil(); 
     miCoche.llenaTanque(); 
-    miCoche.imprimeAutomovil();
-    return 0;
+    miCoche.imprimeAutomovil(); 
+    return 0; 
 }
